fix: refactor contact file
